server/requirements.txt
bcrypt==4.0.1
blinker==1.6.2
click==8.1.7
Flask==2.3.3
Flask-Bcrypt==1.0.1
flask-marshmallow==0.15.0
Flask-SQLAlchemy==3.0.5
greenlet==2.0.2
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2==2.9.7
PyJWT==2.8.0
SQLAlchemy==2.0.20
typing_extensions==4.7.1
Werkzeug==2.3.7
